+++ docs/index.html
@@ -1,1 +1,1 @@
-<!DOCTYPE html><html><head><meta charset="UTF-8"><meta name="viewport" content="width=device-width, initial-scale=1.0"><meta http-equiv="X-UA-Compatible" content="ie=edge"><title>128 UI Style Guide</title></head><body><div id="rsg-root"></div><script src="build/bundle.2407623d.js"></script></body></html>
+<!DOCTYPE html><html><head><meta charset="UTF-8"><meta name="viewport" content="width=device-width, initial-scale=1.0"><meta http-equiv="X-UA-Compatible" content="ie=edge"><title>128 UI Style Guide</title></head><body><div id="rsg-root"></div><script src="build/bundle.1d3f49cd.js"></script></body></html>
