+++ package.json
@@ -1,14 +1,11 @@
 {
   "name": "@128technology/ui",
-  "version": "0.4.1",
+  "version": "0.4.2",
   "description": "128 Technology UI component library.",
   "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "types": "index.d.ts",
   "scripts": {
-    "copyDefs": "node ./copyDefs",
-    "build:dev": "webpack --mode development --config webpack.development.config.js && npm run copyDefs",
-    "build:prod": "webpack --mode production --config webpack.production.config.js && npm run copyDefs",
-    "build": "npm run clean:dist && npm run build:dev && npm run build:prod",
+    "build": "npm run clean:dist && webpack",
     "styleguide": "styleguidist server",
     "styleguide:build": "npm run clean:styleguide && styleguidist build",
     "lint:js": "eslint --ext .js --ext .jsx src/components",
@@ -52,7 +49,6 @@
     "eslint-plugin-prettier": "^2.3.1",
     "eslint-plugin-react": "^7.4.0",
     "file-loader": "^2.0.0",
-    "fs-extra": "^7.0.0",
     "ignore-styles": "^5.0.1",
     "jquery": "3.2.1",
     "jsdom": "^11.5.1",
@@ -75,7 +71,8 @@
     "url-loader": "^1.1.1",
     "webpack": "^4.19.0",
     "webpack-cli": "^3.1.0",
-    "webpack-merge": "^4.1.0"
+    "webpack-merge": "^4.1.0",
+    "webpack-node-externals": "^1.7.2"
   },
   "peerDependencies": {
     "react": "^16.0.0",
@@ -83,7 +80,9 @@
     "@material-ui/core": "^3.1.0"
   },
   "dependencies": {
+    "@babel/cli": "^7.1.2",
     "@mdi/font": "^2.7.94",
+    "@types/react-select": "^2.0.6",
     "classnames": "^2.2.5",
     "immutable": "^4.0.0-rc.9",
     "javascript-natural-sort": "^0.7.1",
