--- src/components/Duration/index.d.ts
@@ -1,9 +1,0 @@
-import * as React from 'react';
-
-export interface IDurationProps {
-  duration?: number;
-}
-
-declare class Duration extends React.Component<IDurationProps, any> {}
-
-export default Duration;
