--- webpack.development.config.js
@@ -1,8 +1,0 @@
-const merge = require('webpack-merge');
-const webpackBase = require('./webpack.base.config');
-
-const webpackConfig = {
-  devtool: 'sourcemap'
-};
-
-module.exports = merge(webpackBase, webpackConfig);
